+++ vite-project/src/components/MenuPage/index.jsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 import { MenuNav } from "./menuNav";
 import { supahShakes } from "./utility/menu/supah-shakes";
 import OpenModalButton from "../OpenModalButton/index";
-import { AddItem } from "./utility/modals/AddItem";
-import { EditItem } from "./utility/modals/EditItem";
+import { AddItem } from "./utility/forms/AddItem";
+import { EditItem } from "./utility/forms/EditItem";
 import { BackCardItem, FrontCardItem, EmptyCardItem } from "./utility/CardShape";
 import { menuCategories } from "./utility/menu/menu-categories";
 import { combinedMenu } from "./utility/menu/combined-menu";
